+++ CHANGES.rst
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+unlreleased
+-------
+
+- [feature] Drop support for python 2.7. From now on, only support python 3.6 and up
+
 1.1.2
 -------